Reorganize profile components into unified structure
