+++ src/components/profile/discipline-card.tsx
