+++ src/components/profile/ratings-chart.tsx
