+++ src/modules/members/ui/views/member-id-view.tsx
@@ -7,7 +7,7 @@ import { useSuspenseQueries } from "@tanstack/react-query";
 import { LoadingState } from "@/components/loading-state";
 import { ErrorState } from "@/components/error-state";
 import { Banner } from "@/components/banner";
-import { Profile } from "@/modules/profile/ui/components/profile";
+import { Profile } from "@/components/profile";
 
 interface MemberIdViewProps {
   userId: string;
